refactor(front): create profile struct
