+++ front/src/App.tsx
@@ -2,9 +2,9 @@ import { Routes, Route } from 'react-router-dom';
 import HomePage from './pages/HomePage';
 import LoginPage from './pages/LoginPage';
 import ProfilePage from './pages/ProfilePage';
-import EditProfile from './pages/EditProfilePage';
-import AdminProfiles from './pages/ManageProfilepage';
-import SettingsProfiles from './pages/ChangeProfileSettings';
+import EditProfile from './pages/profile/EditProfilePage';
+import AdminProfiles from './pages/profile/ManageProfilepage';
+import SettingsProfiles from './pages/profile/ChangeProfileSettings';
 import SelectMoviePage from './pages/SelectMoviePage';
 import Registration from './pages/signup/Registration';
 import SelectPlan from './pages/signup/SelectPlan';
